Fixed issues in db.js and changed .gitignore
In the loadDevice function in db.js, before the result is being used, added an if statement fixing an issue where the code tries to subscript null, which fails. Also, deleted unnecessary code.
In the disconnectDevice function, fixed an issue where the code would try to delete the returned value from the MATCH statement, even if it’s null which fails.
In .gitignore, instead of ignoring the socket.io, neo4j-driver and express directories directly, it ignores the whole node_modules directory but keeps the db module not ignored.