Fix bugs in `src/index.js` and in `device.js`
Moved from using `Device.deviceFromData` to using the regular constructor of the `Device` class in `src/index.js`. Fixed an issue in `src/objects/device.js` where the constructor would set the properties of the device based on the given `options` argument, and then assign undefined values to those properties.